Re-working the cluster synchronization logic, and doing cluster integration tests.
git-svn-id: http://ikube.googlecode.com/svn/trunk@744 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/core/src/main/java/ikube/model/Server.java
@@ -132,7 +132,7 @@ public class Server extends Persistable implements Comparable<Server> {
 		if (this.getId() == 0) {
 			return super.hashCode();
 		}
-		return (int) this.getId();
+		return (int) (long) this.getId();
 	}
 
 	@Override
